move tests for nd binning
